better format for swiftui example
