+++ Example/Sources/UIKit/Components/UIColor+DynamicColor.swift
